+++ sys_win.c
@@ -476,7 +476,7 @@ DWORD countMemUsage()
 
 #define signatureSize	4
 
-void**  mymalloc( long numBytes )					// Memory allocation, use Handles
+void**  mymalloc( size_t numBytes )					// Memory allocation, use Handles
 {
 	LPSTR *mHand;
 	char *p;
